requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.10
langgraph>=0.0.62
chromadb>=0.5.0
python-dotenv>=1.0.1
